+++ nuxeo/models.py
@@ -235,6 +235,7 @@ class Blob(Model):
     """ Blob superclass used for metadata. """
 
     _valid_properties = {
+        "batchId": "",
         "chunkCount": 0,
         "fileIdx": None,
         "mimetype": None,
@@ -258,10 +259,34 @@ class Blob(Model):
             model.uploadedSize = model.size
         return model
 
+    @property
+    def batch_id(self):
+        """Kept for compatibility reasons, will be removed in version 4.0."""
+        from warnings import warn
+
+        warn(
+            "batch_id is deprecated and will be removed in version 4.0. Use batchId instead.",
+            DeprecationWarning,
+            2,
+        )
+        return self.batchId
+
+    @batch_id.setter
+    def batch_id(self, value):
+        """Kept for compatibility reasons, will be removed in version 4.0."""
+        from warnings import warn
+
+        warn(
+            "batch_id is deprecated and will be removed in version 4.0. Use batchId instead.",
+            DeprecationWarning,
+            2,
+        )
+        self.batchId = value
+
     def to_json(self):
         # type: () -> Dict[Text, Text]
         """ Return a JSON object used during the upload. """
-        return {"upload-batch": self.batch_id, "upload-fileId": text(self.fileIdx)}
+        return {"upload-batch": self.batchId, "upload-fileId": text(self.fileIdx)}
 
 
 class BufferBlob(Blob):